feat: HomeView Hi-Fi UI 1차 완료
