+++ Hippo/Shared/Presentation/Home/DisplayModels/OperationAssetDisplayModel.swift
