+++ Hippo/Shared/Presentation/Home/DisplayModels/OperationDisplayModel.swift
